+++ src/modules/layerInformation/components/LayerInformation.vue
@@ -83,7 +83,7 @@ export default {
             return this.downloadLinks !== null;
         },
         showUrl () {
-            if (this.layerInfo.typ === "GROUP" && Array.isArray(this.layerInfo.layers)) {
+            if (this.layerInfo.typ.startsWith("GROUP") && Array.isArray(this.layerInfo.layers)) {
                 const selectedLayer = this.layerInfo.layers[this.selectedOption];
 
                 return selectedLayer?.url && this.showUrlGlobal !== false && this.layerInfo.urlIsVisible !== false;
@@ -91,7 +91,7 @@ export default {
             return this.layerInfo.url && this.layerInfo.typ !== "SensorThings" && this.showUrlGlobal !== false && this.layerInfo.urlIsVisible !== false;
         },
         selectedMetaURLs () {
-            if (this.layerInfo.typ === "GROUP" && Array.isArray(this.layerInfo.layers)) {
+            if (this.layerInfo.typ.startsWith("GROUP") && Array.isArray(this.layerInfo.layers)) {
                 const selectedLayer = this.layerInfo.layers[this.selectedOption],
                     metaID = selectedLayer?.metaID;
 
@@ -109,7 +109,7 @@ export default {
         layerUrl () {
             const layer = this.layerInfo;
 
-            if (layer.typ === "GROUP" && layer.layers) {
+            if (layer.typ.startsWith("GROUP") && layer.layers) {
                 const selectedLayer = layer.layers[this.selectedOption];
 
                 return selectedLayer ? this.getLayerAddress(selectedLayer) : "";
@@ -120,7 +120,7 @@ export default {
             return this.layerInfo.legendURL;
         },
         layerTyp () {
-            if (this.layerInfo.typ !== "GROUP") {
+            if (!this.layerInfo.typ.startsWith("GROUP")) {
                 return `${this.layerInfo.typ}-${this.$t("common:modules.layerInformation.addressSuffix")}`;
             }
 
@@ -163,7 +163,7 @@ export default {
     created () {
         this.setConfigParams(this.configJs);
 
-        if (this.layerInfo.typ === "GROUP") {
+        if (this.layerInfo.typ.startsWith("GROUP")) {
             this.createDropdownOptions();
         }
     },
@@ -323,7 +323,7 @@ export default {
         id="modules-layer-information"
     >
         <div
-            v-if="layerInfo.typ === 'GROUP'"
+            v-if="layerInfo.typ.startsWith('GROUP')"
             class="form-floating mb-3"
         >
             <select